Comprobando que la cantidad de numeros que se puede calcular con int es menor que con long long

+++ numeros largos.c
@@ -3,9 +3,9 @@
 #include <stdint.h>
 
 int main(){
-long long num =0;
-long long num2 = 1;
-long long fibo = 0;
+int num =0;
+int num2 = 1;
+int fibo = 0;
 int max_size;
 printf("Ingrese un limite:\n" );
 scanf("%d",&max_size);
@@ -15,7 +15,7 @@ for ( i = 0; i < max_size; i++)
     fibo = num + num2;
     num = num2;
     num2 = fibo;
-    printf("El numero actual es %I64d \n",fibo );
+    printf("El numero actual es %d \n",fibo );
 }
